pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "cropdoc"
version = "0.1.0"
description = "AI Crop Disease Detector - Identify plant diseases from leaf images"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Mukunda Katta" },
]

dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "pydantic>=2.0",
    "click>=8.0",
    "rich>=13.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
cropdoc = "cropdoc.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
